requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-geoposition==0.3.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
psycopg2-binary==2.8.4
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
